+++ db-queue-scheduler-core/src/main/java/ru/yoomoney/tech/dbqueue/scheduler/internal/schedule/ScheduledTaskExecutionContext.java
@@ -1,5 +1,7 @@
 package ru.yoomoney.tech.dbqueue.scheduler.internal.schedule;
 
+import ru.yoomoney.tech.dbqueue.scheduler.models.ScheduledTaskExecutionResult;
+
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import javax.annotation.concurrent.NotThreadSafe;
@@ -22,6 +24,10 @@ public class ScheduledTaskExecutionContext {
     private Instant lastExecutionStartTime;
     @Nullable
     private Instant lastExecutionFinishTime;
+    @Nullable
+    private ScheduledTaskExecutionResult.Type executionResultType;
+    @Nullable
+    private Long attemptsCount;
 
     public void setLastExecutionStartTime(@Nonnull Instant lastExecutionStartTime) {
         this.lastExecutionStartTime = requireNonNull(lastExecutionStartTime, "lastExecutionStartTime");
@@ -29,6 +35,14 @@ public class ScheduledTaskExecutionContext {
 
     public void setLastExecutionFinishTime(@Nonnull Instant lastExecutionFinishTime) {
         this.lastExecutionFinishTime = requireNonNull(lastExecutionFinishTime, "lastExecutionFinishTime");
+    }
+
+    public void setExecutionResultType(@Nonnull ScheduledTaskExecutionResult.Type executionResultType) {
+        this.executionResultType = requireNonNull(executionResultType, "executionResultType");
+    }
+
+    public void setAttemptsCount(@Nullable Long attemptsCount) {
+        this.attemptsCount = attemptsCount;
     }
 
     @Nonnull
@@ -41,11 +55,23 @@ public class ScheduledTaskExecutionContext {
         return Optional.ofNullable(lastExecutionFinishTime);
     }
 
+    @Nonnull
+    public Optional<ScheduledTaskExecutionResult.Type> getExecutionResultType() {
+        return Optional.ofNullable(executionResultType);
+    }
+
+    @Nonnull
+    public Optional<Long> getAttemptsCount() {
+        return Optional.ofNullable(attemptsCount);
+    }
+
     @Override
     public String toString() {
         return "ScheduledTaskExecutionContext{" +
                 "lastExecutionStartTime=" + lastExecutionStartTime +
                 ", lastExecutionFinishTime=" + lastExecutionFinishTime +
+                ", executionResultType=" + executionResultType +
+                ", attemptsCount=" + attemptsCount +
                 '}';
     }
 }